iterate on config modelling [ci skip]
